+++ src/plugins/multiuserchat/multiuserchatwindow.cpp
@@ -310,8 +310,8 @@ INotification MultiUserChatWindow::notifyMessage(INotifications *ANotifications,
 			{
 				if (Options::node(OPV_NOTIFICATIONS_TABPAGE_SHOWMINIMIZED).value().toBool())
 					page->showMinimizedTabPage();
-				notify.data.insert(NDR_ALERT_WIDGET,(int)page->instance());
-				notify.data.insert(NDR_TABPAGE_OBJECT,(int)page->instance());
+				notify.data.insert(NDR_ALERT_WIDGET,(qint64)page->instance());
+				notify.data.insert(NDR_TABPAGE_OBJECT,(qint64)page->instance());
 				notify.data.insert(NDR_TABPAGE_PRIORITY,TPNP_NEW_MESSAGE);
 				notify.data.insert(NDR_TABPAGE_ICONBLINK,true);
 			}
